cleanup setup, dont drop all data as default

+++ src/Infrastructure/Setup/sql/SetupDatabase.php
@@ -19,19 +19,17 @@ class SetupDatabase {
 
 	}
 
-	public function run():void {
+	public function run(bool $clear = false):void {
 	    global $DIC;
 
+	    if ($clear) {
+            $DIC->database()->dropTable(QuestionEventStoreAr::STORAGE_NAME, false);
+            $DIC->database()->dropTable(QuestionListItemAr::STORAGE_NAME, false);
+            $DIC->database()->dropTable(QuestionAr::STORAGE_NAME, false);
+            $DIC->database()->dropTable(SimpleStoredAnswer::STORAGE_NAME, false);
+            $DIC->database()->dropTable(AssessmentResultEventStoreAr::STORAGE_NAME, false);
+	    }
 
-        $DIC->database()->dropTable(QuestionEventStoreAr::STORAGE_NAME, false);
-        $DIC->database()->dropTable(QuestionListItemAr::STORAGE_NAME, false);
-        $DIC->database()->dropTable(QuestionAr::STORAGE_NAME, false);
-        $DIC->database()->dropTable(SimpleStoredAnswer::STORAGE_NAME, false);
-        $DIC->database()->dropTable("asq_user_answer_scpre", false);
-        $DIC->database()->dropTable("asq_user_test_score", false);
-
-
-        
         QuestionEventStoreAr::updateDB();
 	    QuestionListItemAr::updateDB();
 	    QuestionAr::updateDB();
@@ -44,12 +42,6 @@ class SetupDatabase {
         $DIC->database()->query("UPDATE copg_pobj_def SET component = 'Services/AssessmentQuestion',  class_name = 'AsqPageObject', directory = 'src/UserInterface/Web/Page' where parent_type = 'asqq'");
 
         $this->cleanupContentPages();
-
-        ////////////////////////////////////////////////////
-        /// Test Object Tables
-        $this->changeTestObjectTables();
-        $this->cleanTestObjectTables();
-        ////////////////////////////////////////////////////
 
         echo "Setup wurde durchgefüht. Datentabellen wurden installiert / aktualisiert. ACHTUNG allenfalls muss vorher via setup/setup.php die Ctrl-Struktur neu geladen werden. In diesem Fall dieses Setup erneut ausführen.<br><br>";
         echo "Es müsste nun neben dem Setup / Resetup ASQ ein neuer Tab 'exAsqExamplesGUI' angezeigt werden<br><br>";
@@ -98,42 +90,5 @@ class SetupDatabase {
             "DELETE FROM page_object WHERE parent_type = %s",
             ['text'], ['asqa']
         );
-    }
-
-    protected function changeTestObjectTables()
-    {
-        global $DIC; /* @var \ILIAS\DI\Container $DIC */
-
-        if( $DIC->database()->tableColumnExists('tst_test_question', 'question_uid') )
-        {
-            $DIC->database()->dropTableColumn('tst_test_question', 'question_uid');
-        }
-
-        if( !$DIC->database()->tableColumnExists('tst_test_question', 'question_uid') )
-        {
-            $DIC->database()->addTableColumn('tst_test_question', 'question_uid', array(
-                'type' => 'text',
-                'notnull' => false,
-                'length' => 64,
-                'default' => ''
-            ));
-        }
-
-        if( !$DIC->database()->tableColumnExists('tst_test_question', 'revision_id') )
-        {
-            $DIC->database()->addTableColumn('tst_test_question', 'revision_id', array(
-                'type' => 'text',
-                'notnull' => false,
-                'length' => 64,
-                'default' => ''
-            ));
-        }
-    }
-
-    protected function cleanTestObjectTables()
-    {
-        global $DIC; /* @var \ILIAS\DI\Container $DIC */
-
-        $DIC->database()->manipulate("TRUNCATE TABLE tst_test_question");
     }
 }
